+++ spec/rubocop/cop/lint/string_conversion_in_interpolation_spec.rb
@@ -24,4 +24,9 @@ describe Rubocop::Cop::Lint::StringConversionInInterpolation do
     inspect_source(cop, '"this is the #{result}"')
     expect(cop.offenses).to be_empty
   end
+
+  it 'does not explode on empty interpolation' do
+    inspect_source(cop, '"this is #{} silly"')
+    expect(cop.offenses).to be_empty
+  end
 end