+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 * New cop `GuardClause` checks for conditionals that can be replaced by guard clauses. ([@bbatsov][])
 
+### Bugs fixed
+
+* [#904](https://github.com/bbatsov/rubocop/issues/904) Fixed a NPE in `LiteralInInterpolation`. ([@bbatsov][])
+* [#904](https://github.com/bbatsov/rubocop/issues/904) Fixed a NPE in `StringConversionInInterpolation`. ([@bbatsov][])
+
 ## 0.19.1 (17/03/2014)
 
 ### Bugs fixed